[red-knot] Add boundness and declaredness tests
This changeset adds new tests for public types of symbols and
diagnostics, considering all possible declaredness and boundness states.
This was previously not the case as [discussed here].

Note that this is a mere documentation of the current behavior. There is
still an [open ticket] questioning some of these choices (or unintential
behaviors).

| **Public type**  | declared     | possibly-undeclared        | undeclared   |
| ---------------- | ------------ | -------------------------- | ------------ |
| bound            | `T_declared` | `T_declared \| T_inferred` | `T_inferred` |
| possibly-unbound | `T_declared` | `T_declared \| T_inferred` | `T_inferred` |
| unbound          | `T_declared` | `T_declared`               | `Unknown`    |

| **Diagnostic**   | declared | possibly-undeclared       | undeclared          |
| ---------------- | -------- | ------------------------- | ------------------- |
| bound            |          |                           |                     |
| possibly-unbound |          | `possibly-unbound-import` |                     |
| unbound          |          |                           | `unresolved-import` |

Made sure that the respective test fails if I add the previous case
again in `symbol_by_id`:

```rs
Symbol::Type(inferred_ty, Boundness::Bound) => {
    Symbol::Type(inferred_ty, Boundness::Bound)
}
```

[discussed here]: https://github.com/astral-sh/ruff/pull/15019#discussion_r1890970501
[open ticket]: https://github.com/astral-sh/ruff/issues/14297