Separate grouped and ungrouped nodes more clearly in AST generator
This is a minor cleanup to the AST generation script to make a clearer
separation between nodes that do appear in a group enum, and those that
don't.  There are some types and methods that we create for every
syntax node, and others that refer to the group that the syntax node
belongs to, and which therefore don't make sense for ungrouped nodes.
This new separation makes it clearer which category each definition is
in, since you're either inside of a `for group in ast.groups` loop, or a
`for node in ast.all_nodes` loop.