Remove customizable reference enum names
The AST generator creates a reference enum for each syntax group — an
enum where each variant contains a reference to the relevant syntax
node.  Previously you could customize the name of the reference enum for
a group — primarily because there was an existing `ExpressionRef` type
that wouldn't have lined up with the auto-derived name `ExprRef`.  This
follow-up PR is a simple search/replace to switch over to the
auto-derived name, so that we can remove this customization point.