add DYNAMIC string flag
some cases really seem to need the information from the Generator rather than
the quotes on the StringLiteral itself. It's also hard to tell whether the
StringLiteralFlags have been set intentionally or if they just default to single
quotes. this flag allows callers to indicate explicitly that the flag is
effectively not set